+++ lib/src/loading.dart
@@ -5,4 +5,13 @@ typedef LoadingStateBuilder = Widget Function(BuildContext context);
 
 /// Loading state
 // ignore: prefer_function_declarations_over_variables
-LoadingStateBuilder sduiLoadingState = (context) =>  const CircularProgressIndicator();
+LoadingStateBuilder sduiLoadingState = (context) =>
+    Scaffold(
+      appBar: AppBar(
+        title: const Text('Loading...'),
+        automaticallyImplyLeading: false,
+      ),
+      body: const Center(
+        child: CircularProgressIndicator()
+      )
+    );